+++ nerdlets/geo-ops-nerdlet/styles.scss
@@ -9,7 +9,7 @@
 @import './nrds.scss';
 
 // Main (UI/View) pages
-@import './CreateMap.scss';
+@import './CreateMap/CreateMap.scss';
 @import './ViewMap/ViewMap.scss';
 @import './ViewMap/LocationMetadata.scss';
